[AzureCommunicationFileUploadAdapter.ts] Combine the attachment manager into this adapter so all attachments logic is in one place
